refactor: add `executors` package
