+++ src/speaches/executors/whisper/utils.py
