+++ src/speaches/routers/models.py
@@ -5,14 +5,15 @@ from fastapi import (
     HTTPException,
 )
 
-from speaches import kokoro_utils, piper_utils
 from speaches.api_types import (
     ListModelsResponse,
     Model,
     ModelTask,
 )
+from speaches.executors.kokoro.utils import get_kokoro_models
+from speaches.executors.piper.utils import get_piper_models
+from speaches.executors.whisper.utils import list_local_whisper_models, list_whisper_models
 from speaches.model_aliases import ModelId
-from speaches.whisper_utils import list_local_whisper_models, list_whisper_models
 
 router = APIRouter(tags=["models"])
 
@@ -23,8 +24,8 @@ router = APIRouter(tags=["models"])
 def get_models(task: ModelTask | None = None) -> ListModelsResponse:
     models: list[Model] = []
     if task is None or task == "text-to-speech":
-        models.extend(kokoro_utils.get_kokoro_models())
-        models.extend(piper_utils.get_piper_models())
+        models.extend(get_kokoro_models())
+        models.extend(get_piper_models())
     if task is None or task == "automatic-speech-recognition":
         if os.getenv("HF_HUB_OFFLINE") is not None:
             models.extend(list(list_local_whisper_models()))
@@ -37,8 +38,8 @@ def get_models(task: ModelTask | None = None) -> ListModelsResponse:
 @router.get("/v1/models/{model_id:path}")
 def get_model(model_id: ModelId) -> Model:
     models: list[Model] = []
-    models.extend(kokoro_utils.get_kokoro_models())
-    models.extend(piper_utils.get_piper_models())
+    models.extend(get_kokoro_models())
+    models.extend(get_piper_models())
     if os.getenv("HF_HUB_OFFLINE") is not None:
         models.extend(list(list_local_whisper_models()))
     else:
